Only joining the raster stats thread when used.

+++ src/pygeoprocessing/multiprocessing/raster_calculator.py
@@ -9,6 +9,9 @@ import signal
 import sys
 import time
 
+import numpy
+from osgeo import gdal
+
 from ..geoprocessing import _is_raster_path_band_formatted
 from ..geoprocessing import _LARGEST_ITERBLOCK
 from ..geoprocessing import _LOGGING_PERIOD
@@ -19,8 +22,6 @@ from ..geoprocessing import get_raster_info
 from ..geoprocessing import iterblocks
 from ..geoprocessing import LOGGER
 from ..geoprocessing_core import DEFAULT_GTIFF_CREATION_TUPLE_OPTIONS
-from osgeo import gdal
-import numpy
 
 if sys.version_info >= (3, 8):
     import multiprocessing.shared_memory
@@ -560,16 +561,17 @@ def raster_calculator(
     for _ in range(n_workers):
         block_offset_queue.put(None)
 
-    LOGGER.info('wait for stats worker to complete')
-    stats_worker.join(_MAX_TIMEOUT)
-    if stats_worker.is_alive():
-        LOGGER.error(
-            f'stats worker {stats_worker.pid} '
-            'didn\'t terminate, sending kill signal.')
-        try:
-            os.kill(stats_worker.pid, signal.SIGTERM)
-        except Exception:
-            LOGGER.exception(f'unable to kill {stats_worker.pid}')
+    if calc_raster_stats:
+        LOGGER.info('wait for stats worker to complete')
+        stats_worker.join(_MAX_TIMEOUT)
+        if stats_worker.is_alive():
+            LOGGER.error(
+                f'stats worker {stats_worker.pid} '
+                'didn\'t terminate, sending kill signal.')
+            try:
+                os.kill(stats_worker.pid, signal.SIGTERM)
+            except Exception:
+                LOGGER.exception(f'unable to kill {stats_worker.pid}')
 
     # wait for the workers to join
     LOGGER.info('all work sent, waiting for workers to finish')
